requirements.txt
aiofiles==0.7.0
anyio==3.3.4
appdirs==1.4.3
appier==1.21.7
asgiref==3.3.4
astroid==2.5.6
async-timeout==3.0.1
asyncstdlib==3.9.2
attrs==21.2.0
backcall==0.2.0
bidict==0.21.2
boto3==1.17.72
botocore==1.20.72
bson==0.5.10
CacheControl==0.12.6
cachetools==4.2.2
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==35.0.0
decorator==5.0.9
distlib==0.3.0
distro==1.4.0
Django==3.2.8
django-cors-headers==3.10.0
dnspython==2.1.0
facebook-sdk==3.1.0
google-api==0.1.12
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
h11==0.12.0
html5lib==1.0.1
httplib2==0.19.1
httptools==0.2.0
idna==2.8
ipaddr==2.2.0
ipython==7.23.1
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
jmespath==0.10.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
matplotlib-inline==0.1.2
mccabe==0.6.1
msgpack==0.6.2
multidict==5.1.0
Naked==0.1.31
packaging==20.3
parso==0.8.2
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
progress==1.5
prompt-toolkit==3.0.18
protobuf==3.17.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.11.0
Pygments==2.9.0
PyJWT==2.3.0
pymongo==3.12.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2021.1
PyYAML==6.0
requests==2.22.0
rsa==4.7.2
s3transfer==0.4.2
shellescape==3.8.1
six==1.14.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.16.0
toml==0.10.2
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.12.1
